Split core tests by area
